Don't invoke this.focusInput in moustrap callback
When calling this.focusInput() in the moustrap callback it is possible for this.commandPaletteInput.input to be null since this.commandPaletteInput is set during the render which is not guaranteed to happen first. This creates a race condition that can lead to an error being logged on the console in some cases.

this.focusInput() is still called by afterOpenModal and thus the desired effect is acheived.